move `DbConnReaper` into `database.go`
